Java Installer Version 2.0
This is a significant under the hood change for the java installer. It
involves making an android-style system for passing arguments as well as
a backstack for easily managing where the back button should take. It
also adds automatic detection of both the JRE creator tar and a JRE
created and summed by the installer. At the end of the install process,
it now deletes everything but the final JRE.tar.gz and the md5 file for
the tar.

I also added in gradle to this build. The uberjar target will create a
fat jar suitable for distribution to teams.

The installer now uses the new mDNS format (roborio-<num>-FRC.local).

Change-Id: I8e8fb3582bd2850ea8ee14739fc9022873bb1bd8